Allocator.sub should never raise exceptions
`sub` should have the sense of "subtraction" i.e. a \ b

It shouldn't be an error if a `intersection` b is empty. If a particular
user wishes to impose that constraint, they should do that themselves.
This module should be all about sets of segments (logical or physical)
i.e. sets of integers.

Signed-off-by: David Scott <dave.scott@citrix.com>