Replace `Vg.Make(...).read` with `connect` and `flush`
In future one will `connect` to a volume group and perform read and
write operations against it. If the volume group supports a redo-log
then this will be used to accelerate writes. If a redo-log is in use
then `flush` will cause it to be flushed.

Note: these new functions currently fail with unimplemented errors.

Signed-off-by: David Scott <dave.scott@citrix.com>