Remove `type vg` from Vg.Make().Volume
To make a BLOCK device we don't need the full volume group metadata.
We also don't want to leak a snapshot of the volume group via the
`BLOCK.id` function. Instead, we make `Volume.id` abstract and add a
function `find` to look up LVs by name in the volume group.

Signed-off-by: David Scott <dave.scott@citrix.com>