Use UUIDs to index LVs and for redo log operations
This fixes the failing test added by 6a219e8 for the idempotency of the
LvRename operation.

This preserves the use of name strings for the interface exposed by the Vg
module.

At the moment, lookup by name will be non-performant since we're now storing
the LVs as a Lv.t UuidMap.t instead of a Lv.t StringMap.t. Lookups by name now
have to filter the map on their value rather than keys.

We should add a second index by name and abstract this behind the LVs module.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>