LvTransfer: Update the source with the correct segments
We were updating the source LV with the blocks that we were asking to transfer.
So that an LvTransfer A B s would set the segments of A and B to s. This is
clearly not what we want and was due to the use of `segments` rather than the
new value `segments'`.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>