CA-188133: Use the metadata area properly
We currently do not update the position of the last metadata position
when writing the metadata to disk. This means that if you make
multiple updates without rereading, the new metadata will always be
written in the same area of the disk, thus causing apparent corruption
if you're trying to read at the same time (the 'corruption' is
detected by a CRC check, and is only transient).

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>